test(proxies): 🔄 replace ProxyPlatform with VoidProxy
Replaced the `ProxyPlatform` class with the new `VoidProxy` class in the `ProxiedConnectionTests.cs` and `VoidProxy.cs` files. The functionality of `ProxyPlatform` has been integrated into `VoidProxy`, which maintains the same constructor parameters and methods. Updated the test method `MccConnectsToPaperServerThroughProxy` to utilize `VoidProxy`, reflecting a change in the proxy implementation for the test.
