+++ src/Tests/Integration/Connections/ProxiedConnectionTests.cs
@@ -18,7 +18,7 @@ public class ProxiedConnectionTests : ConnectionTestBase
         using var cancellationTokenSource = new CancellationTokenSource(TimeSpan.FromMinutes(3));
 
         await using var paper = new PaperServer(ExpectedText);
-        await using var proxy = new ProxyPlatform(address: "localhost:25565", port: 25566);
+        await using var proxy = new VoidProxy(address: "localhost:25565", port: 25566);
         await using var mcc = new MinecraftConsoleClient(ExpectedText, address: "localhost:25566");
 
         var proxyTask = proxy.RunAsync(cancellationTokenSource.Token);
